pyproject.toml
[project]
name = "mog-xltpl"
version = "1.0.0"
description = "A Windows-only xltpl fork that preserves VBA, images, and complex Excel formatting using COM"
readme = "README_EN.md"
requires-python = ">=3.8"
dependencies = [
    "xlrd>=1.2.0",
    "xlwt>=1.3.0",
    "openpyxl>=3.1.0",
    "jinja2",
    "six",
    "tomli",
    "pyyaml>=6.0.3",
    "pywin32>=311",
    "pillow>=9.0.0",
]
authors = [
    {name = "mog (forked from Zhang Yu)"}
]
license = {file = "LICENSE"}
keywords = ["Excel", "xls", "xlsx", "spreadsheet", "workbook", "template", "VBA", "COM", "Windows", "preserve", "formatting"]

[project.urls]
"Homepage" = "https://github.com/[YOUR-USERNAME]/mog-xltpl"
"Original" = "https://github.com/zhangyu836/xltpl"
"Bug Tracker" = "https://github.com/[YOUR-USERNAME]/mog-xltpl/issues"

[project.optional-dependencies]
test = [
    "pytest>=7",
]

[project.scripts]
xltpl = "xltpl.cli:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
